Move styles folder into prev
