+++ src/prev/Form/FormNavigation/FormNavigation.js
@@ -8,8 +8,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import globalStyles from '../../../styles/globalStyles';
-import { textStyles } from '../../../styles/theme';
+import globalStyles from '../../styles/globalStyles';
+import { textStyles } from '../../styles/theme';
 import ActionButton from '../../Button/ActionButton';
 import TextLink from '../../Button/TextLink';
 import Heading from '../../Heading/Heading';
